+++ R/plot_3d_scatterbox.R
@@ -27,13 +27,14 @@
 #' @param shapes name of the column with the second categorical factor in a two-way ANOVA design.
 #' @param symsize size of symbols, default set to 3.
 #' @param symthick size of outline of symbol lines (\code{stroke = 1.0}), default set to 1.0.
-#' @param jitter extent of jitter (scatter) of symbols, default is 0.1. Increase to reduce symbol overlap, set to 0 for aligned symbols.  
+#' @param jitter extent of jitter (scatter) of symbols, default is 0.2. Increase to reduce symbol overlap, set to 0 for aligned symbols.  
 #' @param fontsize parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.
 #' @param b_alpha fractional opacity of boxes, default set to 1 (i.e. maximum opacity & zero transparency).
 #' @param s_alpha fractional opacity of symbols, default set to 1 (i.e. maximum opacity & zero transparency).
 #' @param ColSeq logical TRUE or FALSE. Default TRUE for sequential colours from chosen palette. Set to FALSE for distant colours, which will be applied using  \code{scale_fill_grafify2}.
-#' @param ColPal grafify colour palette to apply, default "all_grafify"; alternatives: "okabe_ito", "bright", "pale", "vibrant", "contrast", "muted" "dark", "light".
-#' @param ColRev whether to reverse order of colour choice, default F (FALSE); can be set to T (TRUE).
+#' @param ColPal grafify colour palette to apply, default "okabe_ito"; see \code{\link{graf_palettes}} for available palettes.
+#' @param ColRev whether to reverse order of colour within the selected palette, default F (FALSE); can be set to T (TRUE).
+#' @param SingleColour a colour hexcode (starting with #), a number between 1-154, or names of colours from `grafify` colour palettes to fill along X-axis aesthetic. Accepts any colour other than "black"; use `grey_lin11`, which is almost black.
 #' @param TextXAngle orientation of text on X-axis; default 0 degrees. Change to 45 or 90 to remove overlapping text.
 #' @param ... any additional arguments to pass to \code{ggplot2}[geom_boxplot].
 #'
@@ -44,9 +45,16 @@
 #' @examples
 #' #3d version for 1-way data with blocking
 #' plot_3d_scatterbox(data = data_1w_death, 
-#' xcol = Genotype, ycol = Death, shapes = Experiment)
+#' xcol = Genotype, ycol = Death, 
+#' shapes = Experiment)
 #' #compare above graph to
-#' plot_scatterbox(data = data_1w_death, xcol = Genotype, ycol = Death)
+#' plot_scatterbox(data = data_1w_death, 
+#' xcol = Genotype, ycol = Death)
+#' #single colour graph
+#' plot_3d_scatterbox(data = data_1w_death, 
+#' xcol = Genotype, ycol = Death,
+#' shapes = Experiment,
+#' SingleColour = "pale_grey")
 #' 
 #' #4d version for 2-way data with blocking
 #' plot_4d_scatterbox(data = data_2w_Tdeath, 
@@ -62,32 +70,66 @@
 #' shapes = Block)
 #'
 
-plot_3d_scatterbox <- function(data, xcol, ycol, shapes, symsize = 2.5, symthick = 1.0, jitter = 0.1, fontsize = 20, b_alpha = 1, s_alpha = 1, ColSeq = TRUE, ColPal = "all_grafify", ColRev = FALSE, TextXAngle = 0, ...){
-  P <- ggplot2::ggplot(data, aes(x = factor({{ xcol }}),
-                            y = {{ ycol }},
-                            group = factor({{ xcol }})))+
-    geom_boxplot(aes(fill = factor({{ xcol }})), size = 1,
-                 alpha = {{ b_alpha }},
-                 position = position_dodge(width = 0.8),
-                 width = 0.5,
-                 outlier.alpha = 0, 
-                 ...)+
-    geom_point(size = {{ symsize }}, 
-               stroke = {{ symthick }},
-               alpha = {{ s_alpha }}, colour = "black",
-               position = position_jitterdodge(jitter.width = {{ jitter }},
-                                               dodge.width = 0.8),
-               aes(shape = factor({{ shapes }})))+
-    scale_shape_manual(values = 0:25)+
-    labs(shape = enquo(shapes),
-         fill = enquo(shapes),
-         x = enquo(xcol))+
-    theme_classic(base_size = {{ fontsize }})+
-    theme(strip.background = element_blank())+
-    guides(x = guide_axis(angle = {{ TextXAngle }}))
-  if (ColSeq) {
-    P <- P + scale_fill_grafify(palette = {{ ColPal }}, reverse = {{ ColRev }})
+plot_3d_scatterbox <- function(data, xcol, ycol, shapes, symsize = 2.5, symthick = 1.0, jitter = 0.2, fontsize = 20, b_alpha = 1, s_alpha = 1, ColSeq = TRUE, ColPal = c("okabe_ito", "all_grafify", "bright",  "contrast",  "dark",  "fishy",  "kelly",  "light",  "muted",  "pale",  "r4",  "safe",  "vibrant"), ColRev = FALSE, SingleColour = "NULL", TextXAngle = 0, ...){
+  ColPal <- match.arg(ColPal)
+  if (missing(SingleColour)) {
+    P <- ggplot2::ggplot(data, 
+                         aes(x = factor({{ xcol }}),
+                             y = {{ ycol }},
+                             group = factor({{ xcol }})))+
+      geom_boxplot(aes(fill = factor({{ xcol }})), size = 1,
+                   alpha = {{ b_alpha }},
+                   position = position_dodge(width = 0.8),
+                   width = 0.5,
+                   outlier.alpha = 0, 
+                   ...)+
+      geom_point(size = {{ symsize }}, 
+                 stroke = {{ symthick }},
+                 alpha = {{ s_alpha }}, colour = "black",
+                 position = position_jitterdodge(jitter.width = {{ jitter }},
+                                                 dodge.width = 0.8),
+                 aes(shape = factor({{ shapes }})))+
+      scale_shape_manual(values = 0:25)+
+      labs(x = enquo(xcol),
+           fill = enquo(xcol),
+           shape = enquo(shapes))+
+      theme_classic(base_size = {{ fontsize }})+
+      theme(strip.background = element_blank())+
+      guides(x = guide_axis(angle = {{ TextXAngle }}),
+             fill = guide_legend(order = 1),
+             shape = guide_legend(order = 2))+
+      scale_fill_grafify(palette = {{ ColPal }}, 
+                         reverse = {{ ColRev }}, 
+                         ColSeq = {{ ColSeq }})
   } else {
-    P <- P + scale_fill_grafify2(palette = {{ ColPal }}, reverse = {{ ColRev }})}
+    ifelse(grepl("#", SingleColour), 
+           a <- SingleColour,
+           a <- get_graf_colours({{ SingleColour }}))
+    P <- ggplot2::ggplot(data, 
+                         aes(x = factor({{ xcol }}),
+                             y = {{ ycol }},
+                             group = factor({{ xcol }})))+
+      geom_boxplot(fill = a, 
+                   size = 1,
+                   alpha = {{ b_alpha }},
+                   position = position_dodge(width = 0.8),
+                   width = 0.5,
+                   outlier.alpha = 0, 
+                   ...)+
+      geom_point(size = {{ symsize }}, 
+                 stroke = {{ symthick }},
+                 alpha = {{ s_alpha }}, colour = "black",
+                 position = position_jitterdodge(jitter.width = {{ jitter }},
+                                                 dodge.width = 0.8),
+                 aes(shape = factor({{ shapes }})))+
+      scale_shape_manual(values = 0:25)+
+      labs(x = enquo(xcol),
+           shape = enquo(shapes))+
+      theme_classic(base_size = {{ fontsize }})+
+      theme(strip.background = element_blank())+
+      guides(x = guide_axis(angle = {{ TextXAngle }}),
+             fill = guide_legend(order = 1),
+             shape = guide_legend(order = 2))    
+  }
   P
 }