+++ R/plot_4d_scatterbar.R
@@ -29,14 +29,15 @@
 #' @param symsize size of symbols, default set to 3.
 #' @param symthick size of outline of symbol lines (\code{stroke = 1.0}), default set to 1.0
 #' @param ewid width of error bars, default set to 0.2.
-#' @param jitter extent of jitter (scatter) of symbols, default is 0.1. Increase to reduce symbol overlap, set to 0 for aligned symbols.  
+#' @param jitter extent of jitter (scatter) of symbols, default is 0.2. Increase to reduce symbol overlap, set to 0 for aligned symbols.  
 #' @param fontsize parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.
 #' @param b_alpha fractional opacity of bars, default set to 1 (i.e. maximum opacity & zero transparency).
 #' @param s_alpha fractional opacity of symbols, default set to 1 (i.e. maximum opacity & zero transparency).
 #' @param ColSeq logical TRUE or FALSE. Default TRUE for sequential colours from chosen palette. Set to FALSE for distant colours, which will be applied using  \code{scale_fill_grafify2}.
-#' @param ColPal grafify colour palette to apply, default "all_grafify"; alternatives: "okabe_ito", "bright", "pale", "vibrant", "contrast", "muted" "dark", "light".
-#' @param ColRev whether to reverse order of colour choice, default F (FALSE); can be set to T (TRUE).
+#' @param ColPal grafify colour palette to apply, default "okabe_ito"; see \code{\link{graf_palettes}} for available palettes.
+#' @param ColRev whether to reverse order of colour within the selected palette, default F (FALSE); can be set to T (TRUE).
 #' @param TextXAngle orientation of text on X-axis; default 0 degrees. Change to 45 or 90 to remove overlapping text.
+#' @param ... any additional arguments to pass to \code{ggplot2}[stat_summary] or \code{ggplot2}[geom_point].
 #'
 #' @return This function returns a \code{ggplot2} object of class "gg" and "ggplot".
 #' @export plot_4d_scatterbar
@@ -63,7 +64,8 @@
 #' shapes = Block)
 #'
 
-plot_4d_scatterbar <- function(data, xcol, ycol, bars, shapes, symsize = 2.5, symthick = 1.0, jitter = 0.1, ewid = 0.2, fontsize = 20, b_alpha = 1, s_alpha = 1, ColPal = "all_grafify", ColRev = FALSE, ColSeq = TRUE, TextXAngle = 0){
+plot_4d_scatterbar <- function(data, xcol, ycol, bars, shapes, symsize = 2.5, symthick = 1.0, jitter = 0.2, ewid = 0.2, fontsize = 20, b_alpha = 1, s_alpha = 1, ColPal = c("okabe_ito", "all_grafify", "bright",  "contrast",  "dark",  "fishy",  "kelly",  "light",  "muted",  "pale",  "r4",  "safe",  "vibrant"), ColRev = FALSE, ColSeq = TRUE, TextXAngle = 0, ...){
+  ColPal <- match.arg(ColPal)
   P <- ggplot2::ggplot(data, aes(x = factor({{ xcol }}),
                             y = {{ ycol }},
                             group = interaction(factor({{ bars }}),
@@ -74,32 +76,30 @@ plot_4d_scatterbar <- function(data, xcol, ycol, bars, shapes, symsize = 2.5, sy
                  alpha = {{ b_alpha }}, size = 1,
                  aes(fill = factor({{ bars }})),
                  position = position_dodge(width = 0.8),
-                 fun = "mean")+
+                 fun = "mean", ...)+
     geom_point(size = {{ symsize }}, 
                alpha = {{ s_alpha }}, 
                stroke = {{ symthick }}, 
                colour = "black",
                position = position_jitterdodge(jitter.width = {{ jitter }},
                                                dodge.width = 0.8),
-               aes(shape = factor({{ shapes }})))+
+               aes(shape = factor({{ shapes }})), ...)+
     stat_summary(geom = "errorbar", colour = "black", size = 1, 
                  width = {{ ewid }},
                  fun.data = "mean_sdl",
                  fun.args = list(mult = 1),
-                 position = position_dodge(width = 0.8))+
+                 position = position_dodge(width = 0.8), ...)+
     scale_shape_manual(values = 0:25)+
     labs(fill = enquo(bars),
-         shape = enquo(shapes),
-         x = enquo(xcol))+
+         x = enquo(xcol),
+         shape = enquo(shapes))+
     theme_classic(base_size = {{ fontsize }})+
     theme(strip.background = element_blank())+
-    guides(x = guide_axis(angle = {{ TextXAngle }}))
-  if (ColSeq) {
-    P <- P + scale_fill_grafify(palette = {{ ColPal }}, 
-                                reverse = {{ ColRev }})
-  } else {
-    P <- P + scale_fill_grafify2(palette = {{ ColPal }}, 
-                                 reverse = {{ ColRev }})
-  }
+    guides(x = guide_axis(angle = {{ TextXAngle }}),
+           fill = guide_legend(order = 1),
+           shape = guide_legend(order = 2))+
+    scale_fill_grafify(palette = {{ ColPal }}, 
+                       reverse = {{ ColRev }}, 
+                       ColSeq = {{ ColSeq }})
   P
 }