+++ R/plot_xy_NumGroup.R
@@ -1,6 +1,6 @@
 #' Plot points on a quantitative X - Y plot & a numeric grouping variable.
 #'
-#' This function takes a data table, quantitative X and Y variables, and a numeric grouping variable, and a and plots a graph with using \code{\link[ggplot2]{geom_point}}. The numerical `NumGroup` variable is mapped to the \code{fill} aesthetic of symbols, which receives the `scale_fill_grafify_c` default palette.
+#' This function takes a data table, quantitative X and Y variables, and a numeric grouping variable, and a and plots a graph with using \code{\link[ggplot2]{geom_point}}. The numerical `NumGroup` variable is mapped to the \code{fill} aesthetic of symbols, which receives the `scale_fill_grafify` default quantitative palette (`blue_conti`).
 #' 
 #' This plot is related to \code{\link{plot_xy_CatGroup}} which requires a categorical grouping factor.
 #' When summary statistics (mean/median) are required, use \code{\link{plot_3d_scatterbar}}, \code{\link{plot_3d_scatterbox}} or \code{\link{plot_4d_scatterbox}}. 
@@ -12,8 +12,10 @@
 #' @param symsize size of symbols used by \code{geom_point}. Default set to 2.5, increase/decrease as needed.
 #' @param symthick thickness of symbol border, default set to 1.
 #' @param s_alpha fractional opacity of symbols, default set to 1 (i.e. maximum opacity & zero transparency).
+#' @param ColPal One of five quantitative colour schemes in `grafify`: `blue_conti`, `yellow_conti`, `grey_conti`, `PrGn_div`, `OrBl_div`.
 #' @param TextXAngle orientation of text on X-axis; default 0 degrees. Change to 45 or 90 to remove overlapping text.
 #' @param fontsize parameter of \code{base_size} of fonts in \code{theme_classic}, default set to size 20.
+#' @param ... any additional arguments to pass to \code{ggplot2}[geom_violin].
 #'
 #' @return This function returns a \code{ggplot2} object of class "gg" and "ggplot".
 #' @export plot_xy_NumGroup
@@ -24,17 +26,24 @@
 #' plot_xy_NumGroup(data = mtcars,
 #' xcol = mpg, ycol = disp, NumGroup = cyl,
 #' s_alpha = 0.8)
+#' #change colour palette
+#' plot_xy_NumGroup(data = mtcars,
+#' xcol = mpg, ycol = disp, NumGroup = cyl,
+#' s_alpha = 0.8, 
+#' ColPal = "grey_conti")
 
-plot_xy_NumGroup <- function(data, xcol, ycol, NumGroup, symsize = 2.5, symthick = 1, s_alpha = 1, TextXAngle = 0, fontsize = 20){
+plot_xy_NumGroup <- function(data, xcol, ycol, NumGroup, symsize = 2.5, symthick = 1, s_alpha = 1, ColPal = c("blue_conti", "yellow_conti", "grey_conti", "PrGn_div", "OrBl_div"), TextXAngle = 0, fontsize = 20,...){
+  ColPal <- match.arg(ColPal)
   ggplot2::ggplot(data, aes(x = {{ xcol }},
                             y = {{ ycol }}))+
     geom_point(size = {{ symsize }}, 
                alpha = {{ s_alpha }},
                aes(fill = {{ NumGroup }}),
                shape = 21, 
-               stroke = {{ symthick }})+
+               stroke = {{ symthick }},
+               ...)+
     theme_classic(base_size = {{ fontsize }})+
     theme(strip.background = element_blank())+
     guides(x = guide_axis(angle = {{ TextXAngle }}))+
-    scale_fill_grafify_c()
+    scale_fill_grafify(palette = {{ ColPal }})
 }